Add support for indirect table enumeration
An indirect enumeration is when the last accessed row of a table is
stored in a different table. The `rets` client may request the lookup
of the next row in the first table and the corresponding update of the
row pointer in the second in an atomic operation.

This method is used by spider to fetch e.g the next reserved payment
for an activate operation. Performing the read-next-store cycle in an
atomic step eliminates the risk of concurrently activating the same
payment in multiple transactions.