Allow keeping the LevelDB data on restart
The LevelDB backend stores some meta-data about the tables in ETS. To
persist this information before the backend is shut down the data is
added to the LevelDB table. Upon restart the information is restored
to ETS and the key is removed from LevelDB. This way the meta-data
will neither conflict with regular operations (because it is not in
LevelDB once the backend has started) nor we risk reading back an
outdated copy of it should the backend crash without a chance to
persist the updated ETS table.